[Fix] : CancelProduct 파일 common으로 이동 및 적용
