+++ common/src/main/java/com/takeit/common/application/dto/CancelProduct.java
@@ -1,0 +1,13 @@
+package com.takeit.common.application.dto;
+
+import java.io.Serializable;
+import java.util.UUID;
+
+public record CancelProduct(
+	UUID productId,
+	Long quantity
+) implements Serializable {
+	public static CancelProduct create(UUID productId, Long quantity) {
+		return new CancelProduct(productId, quantity);
+	}
+}
