--- product-service/src/main/java/com/takeit/product/application/dto/product/CancelProduct.java
@@ -1,9 +1,0 @@
-package com.takeit.product.application.dto.product;
-
-import java.util.UUID;
-
-public record CancelProduct(
-        UUID productId,
-        Long quantity
-)  {}
-
