+++ product-service/src/main/java/com/takeit/product/application/service/ProductRedisService.java
@@ -1,8 +1,8 @@
 package com.takeit.product.application.service;
 
+import com.takeit.common.application.dto.CancelProduct;
 import com.takeit.common.exception.CustomException;
 import com.takeit.common.exception.ErrorCode;
-import com.takeit.product.application.dto.product.CancelProduct;
 import com.takeit.product.domain.entity.Product;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -24,14 +24,14 @@ public class ProductRedisService {
     private final RedisTemplate<String, Object> redisTemplate;
 
     private static final String OCCUPY_SCRIPT =
-            "local productKey = KEYS[1] " +
+        "local productKey = KEYS[1] " +
             "local decrementQuantity = tonumber(ARGV[1]) " +
-            "local currentQuantity = tonumber(redis.call('GET', productKey)) " +
-            "if not currentQuantity or currentQuantity < decrementQuantity then " +
+            "local currentStock = tonumber(redis.call('HGET', productKey, 'stock')) " +
+            "if not currentStock or currentStock < decrementQuantity then " +
             "   return -1 " +  // 재고 부족
             "end " +
-            "redis.call('DECRBY', productKey, decrementQuantity) " +
-            "return currentQuantity - decrementQuantity";
+            "redis.call('HINCRBY', productKey, 'stock', -decrementQuantity) " +
+            "return currentStock - decrementQuantity";
     public void saveProduct(Product product) {
         String key = "product:" + product.getUuid();
         Map<String, Object> productMap = new HashMap<>();
@@ -94,7 +94,7 @@ public class ProductRedisService {
     public void cancelProduct(CancelProduct request) {
         String key = "product:" + request.productId();
         log.info("[cancel product] id : {}, amount : {}", request.productId(), request.quantity());
-        redisTemplate.opsForValue().increment(key, request.quantity());
+        redisTemplate.opsForHash().increment(key, "stock",request.quantity());
     }
 
     public void occupyProduct(UUID productId, int quantity) {
